@@ -1,4 +1,4 @@
-name: in_app_browser
+name: flutter_in_app_browser
 description: "A lightweight, customizable in-app browser for Flutter. This package provides a native-feeling, Instagram-style browser experience with smooth animations, gesture controls, and a modern UI—all without external dependencies."
 version: 0.0.1
 homepage: https://nocorps.org/
@@ -7,50 +7,24 @@ issue_tracker: https://github.com/KANAGARAJ-M/in_app_browser/issues
 
 environment:
   sdk: ^3.8.0
-  flutter: ">=1.17.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  url_launcher: 
+  share_plus: 
+  clipboard: 
+  mockito: ^5.4.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^5.0.0
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
-
-  # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/to/asset-from-package
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/to/resolution-aware-images
-
-  # To add custom fonts to your package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/to/font-from-package
+  plugin:
+    platforms:
+      android:
+        package: com.nocorps.flutter_in_app_browser
+        pluginClass: InAppBrowserPlugin
