@@ -7,24 +7,24 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:example/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
+  testWidgets('In-App Browser example app loads correctly', (WidgetTester tester) async {
+    // Build our app and trigger a frame
     await tester.pumpWidget(const MyApp());
 
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
-
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
-
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
+    // Verify the URL input field exists
+    expect(find.byType(TextField), findsOneWidget);
+    
+    // Verify the Open URL button exists
+    expect(find.text('Open URL'), findsOneWidget);
+    
+    // Verify the embedded browser demo button exists
+    expect(find.text('Try Embedded Browser'), findsOneWidget);
+    
+    // Test tapping the button (this would normally open a browser, but in tests it won't actually render)
+    await tester.tap(find.text('Open URL'));
+    await tester.pumpAndSettle();
   });
 }
