@@ -1,6 +1,6 @@
 library;
 
-export 'src/in_app_browser_controller.dart';
 export 'src/in_app_browser_widget.dart';
+export 'src/in_app_browser_controller.dart';
 export 'src/in_app_browser_settings.dart';
 
