@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'dart:async';
+import 'package:clipboard/clipboard.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:share_plus/share_plus.dart';
 import 'in_app_browser_controller.dart';
 import 'in_app_browser_settings.dart';
 import 'platform_view/platform_view_factory.dart';
@@ -10,27 +11,28 @@ class InAppBrowser extends StatefulWidget {
   final String? title;
   final Color? backgroundColor;
   final Color? foregroundColor;
-  final bool showControls;
   final bool showProgressBar;
   final bool enableShare;
   final bool enableRefresh;
   final bool enableBackForward;
+  final bool showControls;
   final VoidCallback? onClosed;
-  
+
   const InAppBrowser({
     super.key,
     required this.initialUrl,
     this.title,
     this.backgroundColor,
     this.foregroundColor,
-    this.showControls = true,
     this.showProgressBar = true,
     this.enableShare = true,
     this.enableRefresh = true,
     this.enableBackForward = true,
+    this.showControls = true,
     this.onClosed,
   });
 
+  /// Opens the browser as a modal bottom sheet
   static Future<void> open(
     BuildContext context,
     String url, {
@@ -41,34 +43,21 @@ class InAppBrowser extends StatefulWidget {
     bool enableShare = true,
     bool enableRefresh = true,
     bool enableBackForward = true,
-  }) async {
-    await Navigator.of(context).push(
-      PageRouteBuilder(
-        pageBuilder: (context, animation, secondaryAnimation) => InAppBrowser(
-          initialUrl: url,
-          title: title,
-          backgroundColor: backgroundColor,
-          foregroundColor: foregroundColor,
-          showProgressBar: showProgressBar,
-          enableShare: enableShare,
-          enableRefresh: enableRefresh,
-          enableBackForward: enableBackForward,
-        ),
-        transitionsBuilder: (context, animation, secondaryAnimation, child) {
-          const begin = Offset(0.0, 1.0);
-          const end = Offset.zero;
-          const curve = Curves.easeInOut;
-
-          var tween = Tween(begin: begin, end: end).chain(
-            CurveTween(curve: curve),
-          );
-
-          return SlideTransition(
-            position: animation.drive(tween),
-            child: child,
-          );
-        },
-        transitionDuration: const Duration(milliseconds: 300),
+  }) {
+    return showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      backgroundColor: Colors.transparent,
+      builder: (context) => InAppBrowser(
+        initialUrl: url,
+        title: title,
+        backgroundColor: backgroundColor,
+        foregroundColor: foregroundColor,
+        showProgressBar: showProgressBar,
+        enableShare: enableShare,
+        enableRefresh: enableRefresh,
+        enableBackForward: enableBackForward,
+        onClosed: () => Navigator.of(context).pop(),
       ),
     );
   }
@@ -77,101 +66,53 @@ class InAppBrowser extends StatefulWidget {
   State<InAppBrowser> createState() => _InAppBrowserState();
 }
 
-class _InAppBrowserState extends State<InAppBrowser> with TickerProviderStateMixin {
-  final InAppBrowserController _controller = InAppBrowserController();
+class _InAppBrowserState extends State<InAppBrowser> {
+  late InAppBrowserController _controller;
   String _currentUrl = '';
-  String _pageTitle = '';
-  bool _isLoading = true;
-  double _progress = 0.0;
+  String _title = '';
+  double _progress = 0;
+  bool _isSecure = false;
   bool _canGoBack = false;
   bool _canGoForward = false;
-  bool _isSecure = false;
-  
-  // Animation controllers
-  late AnimationController _toolbarAnimationController;
-  late AnimationController _progressAnimationController;
-  late Animation<double> _toolbarAnimation;
-  
-  // Gesture tracking
-  bool _isDragging = false;
-  double _dragDistance = 0.0;
-  double _startDragY = 0.0;
-  static const double _dismissThreshold = 150.0;
-  static const double _dragSensitivity = 20.0;
-  
+
   @override
   void initState() {
     super.initState();
     _currentUrl = widget.initialUrl;
-    _pageTitle = widget.title ?? 'Loading...';
-    
-    // Initialize animation controllers
-    _toolbarAnimationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 200),
-    );
-    
-    _progressAnimationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 300),
-    );
-    
-    _toolbarAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _toolbarAnimationController, curve: Curves.easeInOut),
-    );
-    
-    // Setup listeners
-    _setupListeners();
-    
-    // Start animations
-    _toolbarAnimationController.forward();
-    
-    // Load the URL
-    _loadUrl();
+    _title = widget.title ?? '';
   }
-  
-  void _setupListeners() {
-    _controller.onUrlChanged.listen((url) {
+
+  void _onWebViewCreated(dynamic controller) {
+    _controller = InAppBrowserController(controller);
+    _controller.loadUrl(widget.initialUrl);
+    
+    _controller.onPageStarted = (url) {
+      setState(() {
+        _currentUrl = url;
+      });
+    };
+    
+    _controller.onPageFinished = (url) {
       setState(() {
         _currentUrl = url;
         _isSecure = url.startsWith('https://');
       });
-    });
+      _updateNavigationState();
+    };
     
-    _controller.onTitleChanged.listen((title) {
-      setState(() {
-        _pageTitle = title.isNotEmpty ? title : _extractDomainFromUrl(_currentUrl);
-      });
-    });
-    
-    _controller.onProgressChanged.listen((progress) {
+    _controller.onProgressChanged = (progress) {
       setState(() {
         _progress = progress;
       });
-      
-      if (progress >= 1.0) {
-        _progressAnimationController.reverse();
-      } else {
-        _progressAnimationController.forward();
-      }
-    });
+    };
     
-    _controller.onLoadingStateChanged.listen((isLoading) {
+    _controller.onTitleChanged = (title) {
       setState(() {
-        _isLoading = isLoading;
+        _title = title.isNotEmpty ? title : (widget.title ?? '');
       });
-      _updateNavigationState();
-    });
+    };
   }
-  
-  Future<void> _loadUrl() async {
-    try {
-      await _controller.loadUrl(widget.initialUrl);
-    } catch (e) {
-      debugPrint('Failed to load URL: $e');
-    }
-  }
-  
+
   Future<void> _updateNavigationState() async {
     final canGoBack = await _controller.canGoBack();
     final canGoForward = await _controller.canGoForward();
@@ -183,389 +124,253 @@ class _InAppBrowserState extends State<InAppBrowser> with TickerProviderStateMix
       });
     }
   }
+
+  void _onBackPressed() async {
+    if (await _controller.canGoBack()) {
+      _controller.goBack();
+    } else {
+      _onCloseBrowser();
+    }
+  }
+
+  void _onForwardPressed() {
+    _controller.goForward();
+  }
+
+  void _onReloadPressed() {
+    _controller.reload();
+  }
+
+  void _onCloseBrowser() {
+    if (widget.onClosed != null) {
+      widget.onClosed!();
+    }
+  }
   
-  @override
-  void dispose() {
-    _controller.dispose();
-    _toolbarAnimationController.dispose();
-    _progressAnimationController.dispose();
-    super.dispose();
+  void _copyUrl() async {
+    // Copy current URL to clipboard
+    // You'll need to add clipboard package dependency
+    try {
+      await FlutterClipboard.copy(_currentUrl);
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(content: Text('URL copied to clipboard')),
+        );
+      }
+    } catch (e) {
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Failed to copy: $e')),
+        );
+      }
+    }
+  }
+  
+  void _openInExternalBrowser() async {
+    // Open URL in external browser
+    // You'll need to add url_launcher package dependency
+    final url = Uri.parse(_currentUrl);
+    try {
+      if (await canLaunchUrl(url)) {
+        await launchUrl(url, mode: LaunchMode.externalApplication);
+      } else {
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(content: Text('Could not open $url')),
+          );
+        }
+      }
+    } catch (e) {
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Failed to open browser: $e')),
+        );
+      }
+    }
+  }
+
+  void _shareUrl() async {
+    // Share current URL
+    // You'll need to add share_plus package dependency
+    try {
+      await Share.share(_currentUrl, subject: 'Check out this link');
+    } catch (e) {
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Failed to share: $e')),
+        );
+      }
+    }
   }
   
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
     final backgroundColor = widget.backgroundColor ?? theme.scaffoldBackgroundColor;
-    final foregroundColor = widget.foregroundColor ?? theme.primaryColor;
+    final foregroundColor = widget.foregroundColor ?? theme.colorScheme.primary;
     
-    return Scaffold(
-      backgroundColor: backgroundColor,
-      body: Stack(
-        children: [
-          // Main WebView Container
-          Container(
-            margin: EdgeInsets.only(
-              top: MediaQuery.of(context).padding.top + (widget.showControls ? 100 : 0),
-            ),
-            child: Column(
-              children: [
-                // Progress bar
-                if (widget.showProgressBar && _isLoading)
-                  LinearProgressIndicator(
-                    value: _progress,
-                    backgroundColor: Colors.grey[300],
-                    valueColor: AlwaysStoppedAnimation<Color>(foregroundColor),
-                  ),
-                
-                // WebView
-                Expanded(
-                  child: Transform.translate(
-                    offset: Offset(0, _dragDistance * 0.3),
-                    child: Container(
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(_dragDistance > 0 ? 12 : 0),
-                        boxShadow: _dragDistance > 0
-                            ? [
-                                BoxShadow(
-                                  color: Colors.black.withOpacity(0.2),
-                                  blurRadius: 10,
-                                  offset: const Offset(0, 5),
-                                ),
-                              ]
-                            : null,
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      child: PlatformViewFactory.createPlatformView(
-                        initialUrl: widget.initialUrl,
-                        settings: InAppBrowserSettings(),
-                        onWebViewCreated: (controller) {
-                          // WebView was created, ready to use
-                        },
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          
-          // Toolbar
-          if (widget.showControls)
-            AnimatedBuilder(
-              animation: _toolbarAnimation,
-              builder: (context, child) => Transform.translate(
-                offset: Offset(0, (1 - _toolbarAnimation.value) * -100),
-                child: GestureDetector(
-                  onPanStart: (details) {
-                    _isDragging = true;
-                    _dragDistance = 0.0;
-                    _startDragY = details.globalPosition.dy;
-                  },
-                  onPanUpdate: (details) {
-                    if (_isDragging) {
-                      double deltaY = details.globalPosition.dy - _startDragY;
-                      
-                      if (deltaY > _dragSensitivity) {
-                        setState(() {
-                          _dragDistance = deltaY - _dragSensitivity;
-                        });
-                        
-                        if (_dragDistance > 50) {
-                          _toolbarAnimationController.reverse();
-                        }
-                      }
-                    }
-                  },
-                  onPanEnd: (details) {
-                    _isDragging = false;
-                    
-                    if (_dragDistance > _dismissThreshold) {
-                      Navigator.of(context).pop();
-                      widget.onClosed?.call();
-                    } else {
-                      _toolbarAnimationController.forward();
-                      setState(() {
-                        _dragDistance = 0.0;
-                      });
-                    }
-                  },
-                  child: _buildTopToolbar(context, foregroundColor),
-                ),
-              ),
-            ),
-          
-          // Drag indicator
-          if (_dragDistance > 20)
-            Positioned(
-              top: MediaQuery.of(context).padding.top + 30,
-              left: 0,
-              right: 0,
-              child: Center(
-                child: Container(
-                  width: 40,
-                  height: 4,
-                  decoration: BoxDecoration(
-                    color: Colors.grey[400],
-                    borderRadius: BorderRadius.circular(2),
-                  ),
-                ),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-  
-  Widget _buildTopToolbar(BuildContext context, Color foregroundColor) {
     return Container(
-      padding: EdgeInsets.only(
-        top: MediaQuery.of(context).padding.top,
-        left: 16,
-        right: 16,
-        bottom: 1,
-      ),
+      height: MediaQuery.of(context).size.height * 0.9,
       decoration: BoxDecoration(
-        color: Theme.of(context).scaffoldBackgroundColor,
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 4,
-            offset: const Offset(0, 2),
-          ),
-        ],
+        color: backgroundColor,
+        borderRadius: const BorderRadius.vertical(top: Radius.circular(16)),
       ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          // Top row with close and actions
-          Row(
-            children: [
-              // Close button
-              IconButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                  widget.onClosed?.call();
-                },
-                icon: const Icon(Icons.close),
-                style: IconButton.styleFrom(
-                  foregroundColor: foregroundColor,
+      child: ClipRRect(
+        borderRadius: const BorderRadius.vertical(top: Radius.circular(16)),
+        child: Scaffold(
+          backgroundColor: backgroundColor,
+          appBar: AppBar(
+            backgroundColor: backgroundColor,
+            foregroundColor: foregroundColor,
+            elevation: 0,
+            leading: IconButton(
+              icon: const Icon(Icons.close),
+              onPressed: _onCloseBrowser,
+            ),
+            title: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  _title.isNotEmpty ? _title : _currentUrl,
+                  style: const TextStyle(fontSize: 16),
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
                 ),
-              ),
-              
-              // Title and URL
-              Expanded(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
+                Row(
                   children: [
-                    // Title with navigation buttons
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        // Back button
-                        if (widget.enableBackForward)
-                          IconButton(
-                            onPressed: _canGoBack ? () => _controller.goBack() : null,
-                            icon: const Icon(Icons.arrow_back_ios),
-                            iconSize: 16,
-                            style: IconButton.styleFrom(
-                              foregroundColor: _canGoBack ? foregroundColor : Colors.grey,
-                              minimumSize: const Size(32, 32),
-                              padding: const EdgeInsets.all(4),
-                            ),
-                          ),
-                        
-                        // Title (flexible to take remaining space)
-                        Expanded(
-                          child: Text(
-                            _pageTitle,
-                            style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                              fontWeight: FontWeight.bold,
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                            textAlign: TextAlign.center,
-                          ),
+                    if (_isSecure) 
+                      const Icon(Icons.lock, size: 12, color: Colors.green),
+                    const SizedBox(width: 4),
+                    Expanded(
+                      child: Text(
+                        _currentUrl,
+                        style: TextStyle(
+                          fontSize: 12,
+                          color: foregroundColor.withOpacity(0.7),
                         ),
-                        
-                        // Forward button
-                        if (widget.enableBackForward)
-                          IconButton(
-                            onPressed: _canGoForward ? () => _controller.goForward() : null,
-                            icon: const Icon(Icons.arrow_forward_ios),
-                            iconSize: 16,
-                            style: IconButton.styleFrom(
-                              foregroundColor: _canGoForward ? foregroundColor : Colors.grey,
-                              minimumSize: const Size(32, 32),
-                              padding: const EdgeInsets.all(4),
-                            ),
-                          ),
-                      ],
-                    ),
-                    
-                    // URL row with refresh button
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        // Refresh button
-                        if (widget.enableRefresh)
-                          IconButton(
-                            onPressed: () => _controller.reload(),
-                            icon: _isLoading 
-                                ? SizedBox(
-                                    width: 12,
-                                    height: 12,
-                                    child: CircularProgressIndicator(
-                                      strokeWidth: 1.5,
-                                      valueColor: AlwaysStoppedAnimation<Color>(foregroundColor),
-                                    ),
-                                  )
-                                : const Icon(Icons.refresh),
-                            iconSize: 14,
-                            style: IconButton.styleFrom(
-                              foregroundColor: foregroundColor,
-                              minimumSize: const Size(28, 28),
-                              padding: const EdgeInsets.all(2),
-                            ),
-                          ),
-                        
-                        // Security indicator
-                        if (_isSecure) ...[
-                          Icon(
-                            Icons.lock,
-                            size: 12,
-                            color: Colors.green[600],
-                          ),
-                          const SizedBox(width: 4),
-                        ],
-                        
-                        // Domain URL
-                        Expanded(
-                          child: Text(
-                            _extractDomainFromUrl(_currentUrl),
-                            style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                              color: Colors.grey[600],
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                            textAlign: TextAlign.center,
-                          ),
-                        ),
-                        
-                        // Spacer to balance the refresh button
-                        if (widget.enableRefresh)
-                          const SizedBox(width: 28),
-                      ],
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
+                      ),
                     ),
                   ],
                 ),
-              ),
-              
-              // More options
+              ],
+            ),
+            actions: [
               PopupMenuButton<String>(
-                onSelected: _handleMenuAction,
+                icon: const Icon(Icons.more_vert),
+                onSelected: (value) {
+                  switch (value) {
+                    case 'copy':
+                      _copyUrl();
+                      break;
+                    case 'open':
+                      _openInExternalBrowser();
+                      break;
+                    case 'share':
+                      _shareUrl();
+                      break;
+                  }
+                },
                 itemBuilder: (context) => [
-                  if (widget.enableRefresh)
-                    const PopupMenuItem(
-                      value: 'refresh',
-                      child: Row(
-                        children: [
-                          Icon(Icons.refresh),
-                          SizedBox(width: 8),
-                          Text('Refresh'),
-                        ],
-                      ),
+                  const PopupMenuItem(
+                    value: 'copy',
+                    child: Row(
+                      children: [
+                        Icon(Icons.copy, size: 20),
+                        SizedBox(width: 8),
+                        Text('Copy link'),
+                      ],
                     ),
+                  ),
+                  const PopupMenuItem(
+                    value: 'open',
+                    child: Row(
+                      children: [
+                        Icon(Icons.open_in_browser, size: 20),
+                        SizedBox(width: 8),
+                        Text('Open in browser'),
+                      ],
+                    ),
+                  ),
                   if (widget.enableShare)
                     const PopupMenuItem(
                       value: 'share',
                       child: Row(
                         children: [
-                          Icon(Icons.share),
+                          Icon(Icons.share, size: 20),
                           SizedBox(width: 8),
                           Text('Share'),
                         ],
                       ),
                     ),
-                  const PopupMenuItem(
-                    value: 'copy',
-                    child: Row(
-                      children: [
-                        Icon(Icons.copy),
-                        SizedBox(width: 8),
-                        Text('Copy Link'),
-                      ],
-                    ),
-                  ),
-                  const PopupMenuItem(
-                    value: 'external',
-                    child: Row(
-                      children: [
-                        Icon(Icons.open_in_browser),
-                        SizedBox(width: 8),
-                        Text('Open in Browser'),
-                      ],
-                    ),
-                  ),
                 ],
-                icon: Icon(
-                  Icons.more_vert,
-                  color: foregroundColor,
-                ),
               ),
             ],
           ),
-        ],
+          body: Column(
+            children: [
+              if (widget.showProgressBar && _progress < 1.0)
+                LinearProgressIndicator(
+                  value: _progress,
+                  backgroundColor: backgroundColor,
+                  valueColor: AlwaysStoppedAnimation<Color>(foregroundColor),
+                ),
+              Expanded(
+                child: PlatformViewFactory.createPlatformView(
+                  initialUrl: widget.initialUrl,
+                  settings: InAppBrowserSettings(
+                    javascriptEnabled: true,
+                    domStorageEnabled: true,
+                    databaseEnabled: true,
+                    useWideViewPort: true,
+                    allowFileAccess: true,
+                    allowContentAccess: true,
+                    loadWithOverviewMode: true,
+                  ),
+                  onWebViewCreated: _onWebViewCreated,
+                ),
+              ),
+              if (widget.showControls)
+                Container(
+                  padding: const EdgeInsets.symmetric(vertical: 8),
+                  decoration: BoxDecoration(
+                    color: backgroundColor,
+                    border: Border(
+                      top: BorderSide(
+                        color: Colors.grey.withOpacity(0.2),
+                        width: 1,
+                      ),
+                    ),
+                  ),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    children: [
+                      if (widget.enableBackForward)
+                        IconButton(
+                          icon: const Icon(Icons.arrow_back_ios),
+                          onPressed: _canGoBack ? _onBackPressed : null,
+                          color: _canGoBack ? foregroundColor : Colors.grey,
+                        ),
+                      if (widget.enableBackForward)
+                        IconButton(
+                          icon: const Icon(Icons.arrow_forward_ios),
+                          onPressed: _canGoForward ? _onForwardPressed : null,
+                          color: _canGoForward ? foregroundColor : Colors.grey,
+                        ),
+                      if (widget.enableRefresh)
+                        IconButton(
+                          icon: const Icon(Icons.refresh),
+                          onPressed: _onReloadPressed,
+                          color: foregroundColor,
+                        ),
+                    ],
+                  ),
+                ),
+            ],
+          ),
+        ),
       ),
     );
-  }
-  
-  void _handleMenuAction(String action) async {
-    switch (action) {
-      case 'refresh':
-        _controller.reload();
-        break;
-      case 'share':
-        _shareUrl();
-        break;
-      case 'copy':
-        _copyUrl();
-        break;
-      case 'external':
-        _openInExternalBrowser();
-        break;
-    }
-  }
-  
-  void _shareUrl() {
-    // Native sharing would require platform channel implementation
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('Share: $_currentUrl')),
-    );
-  }
-  
-  void _copyUrl() {
-    Clipboard.setData(ClipboardData(text: _currentUrl));
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('URL copied to clipboard')),
-    );
-  }
-  
-  void _openInExternalBrowser() async {
-    // This would be implemented via platform channels
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('Opening in external browser')),
-    );
-  }
-  
-  String _extractDomainFromUrl(String url) {
-    try {
-      final uri = Uri.parse(url);
-      return uri.host.isNotEmpty ? uri.host : url;
-    } catch (e) {
-      if (url.length > 50) {
-        return '${url.substring(0, 47)}...';
-      }
-      return url;
-    }
   }
 }
